+++ app/support/DbAdapter/timelines-posts.js
@@ -7,7 +7,7 @@ import { List } from '../open-lists';
 import { COMMENT_FIELDS, initCommentObject } from './comments';
 import { ATTACHMENT_FIELDS, initAttachmentObject } from './attachments';
 import { POST_FIELDS, initPostObject } from './posts';
-import { sqlIn, sqlIntarrayIn, sqlNotIn, andJoin, orJoin } from './utils';
+import { sqlIn, sqlIntarrayIn, sqlNotIn, andJoin, orJoin, sqlNot } from './utils';
 
 ///////////////////////////////////////////////////
 // Posts Select to fill timeline
@@ -33,29 +33,7 @@ const timelinesPostsTrait = (superClass) =>
     }) {
       withLocalBumps = withLocalBumps && !!viewerId && sort === 'bumped';
 
-      const [
-        // Private feeds viewer can read
-        visiblePrivateFeedIntIds,
-        // Users who banned viewer or banned by viewer (viewer should not see their posts)
-        bannedUsersIds,
-      ] = await Promise.all([
-        viewerId ? this.getVisiblePrivateFeedIntIds(viewerId) : [],
-        viewerId ? this.getUsersBansOrWasBannedBy(viewerId) : [],
-      ]);
-
-      const restrictionsSQL = andJoin([
-        // Privacy
-        viewerId
-          ? orJoin([
-              'not p.is_private',
-              sqlIntarrayIn('p.destination_feed_ids', visiblePrivateFeedIntIds),
-            ])
-          : 'not p.is_protected',
-        // Bans
-        sqlNotIn('p.user_id', bannedUsersIds),
-        // Gone post's authors
-        'u.gone_status is null',
-      ]);
+      const restrictionsSQL = await this.postsVisibilitySQL(viewerId);
 
       /**
        * PostgreSQL is not very good dealing with queries like
@@ -346,25 +324,23 @@ const timelinesPostsTrait = (superClass) =>
         uniqPostsIds,
       );
 
-      const [bannedUsersIds, friendsIds, postsData, attData, { rows: destData }] =
-        await Promise.all([
-          viewerId ? this.getUserBansIds(viewerId) : [],
-          viewerId ? this.getUserFriendIds(viewerId) : [],
-          this.database
-            .select('a.old_url as friendfeed_url', ...postFields)
-            .from('posts as p')
-            .leftJoin('archive_post_names as a', 'p.uid', 'a.post_id')
-            .whereIn('p.uid', uniqPostsIds),
-          this.database
-            .select(...attFields)
-            .from('attachments')
-            .orderBy('ord', 'asc')
-            .orderBy('created_at', 'asc')
-            .whereIn('post_id', uniqPostsIds),
-          this.database.raw(destinationsSQL),
-        ]);
+      const [friendsIds, postsData, attData, destData] = await Promise.all([
+        viewerId ? this.getUserFriendIds(viewerId) : [],
+        this.database
+          .select('a.old_url as friendfeed_url', ...postFields)
+          .from('posts as p')
+          .leftJoin('archive_post_names as a', 'p.uid', 'a.post_id')
+          .whereIn('p.uid', uniqPostsIds),
+        this.database
+          .select(...attFields)
+          .from('attachments')
+          .orderBy('ord', 'asc')
+          .orderBy('created_at', 'asc')
+          .whereIn('post_id', uniqPostsIds),
+        this.database.getAll(destinationsSQL),
+      ]);
 
-      const nobodyIsBanned = bannedUsersIds.length === 0;
+      const notBannedSQLFabric = await this.notBannedActionsSQLFabric(viewerId);
 
       const allLikesSQL = `
       select
@@ -378,8 +354,9 @@ const timelinesPostsTrait = (superClass) =>
         count(*) over (partition by l.post_id) 
       from likes l
           join users u on l.user_id = u.uid
+          join posts p on l.post_id = p.uid
       where ${sqlIn('l.post_id', uniqPostsIds)} 
-        and ${sqlNotIn('l.user_id', bannedUsersIds)}
+        and ${notBannedSQLFabric('l')}
         and u.gone_status is null`;
 
       const foldLikesSql = params.foldLikes
@@ -396,45 +373,40 @@ const timelinesPostsTrait = (superClass) =>
       group by post_id, count 
     `;
 
-      // Don't show comments that viewer don't want to see
-      let hideCommentsSQL = 'true';
-
-      if (params.hiddenCommentTypes.length > 0) {
-        if (params.hiddenCommentTypes.includes(Comment.HIDDEN_BANNED) && !nobodyIsBanned) {
-          hideCommentsSQL = sqlNotIn('user_id', bannedUsersIds);
-        }
-
-        const ht = params.hiddenCommentTypes.filter(
-          (t) => t !== Comment.HIDDEN_BANNED && t !== Comment.VISIBLE,
-        );
-
-        if (ht.length > 0) {
-          hideCommentsSQL += ` and ${(sqlNotIn('hide_type'), ht)}`;
-        }
-      }
-
       const viewerIntId = viewerId ? await this._getUserIntIdByUUID(viewerId) : null;
 
+      const excludeBannedComments = params.hiddenCommentTypes.includes(Comment.HIDDEN_BANNED);
+      const otherExcludedTypes = params.hiddenCommentTypes.filter(
+        (t) => t !== Comment.HIDDEN_BANNED && t !== Comment.VISIBLE,
+      );
+
+      const bannedCommentsSQL = sqlNot(notBannedSQLFabric('c'));
+
+      const commentFilterSQL = andJoin([
+        sqlIn('c.post_id', uniqPostsIds),
+        excludeBannedComments ? sqlNot(bannedCommentsSQL) : 'true',
+        sqlNotIn('c.hide_type', otherExcludedTypes),
+      ]);
+
       const allCommentsSQL = pgFormat(
-        `
-      select
-        ${commentFields.join(', ')},
-        rank() over (partition by post_id order by created_at, id),
-        count(*) over (partition by post_id),
+        `select
+        ${commentFields.map((f) => `c.${f}`).join(', ')},
+        ${bannedCommentsSQL} as hide_as_banned,
+        rank() over (partition by c.post_id order by c.created_at, c.id),
+        count(*) over (partition by c.post_id),
         (select coalesce(count(*), 0) from 
           comment_likes cl
           join users u on cl.user_id = u.id
-          where cl.comment_id = comments.id
-            and cl.user_id not in (select id from users where ${sqlIn('uid', bannedUsersIds)})
+          where cl.comment_id = c.id
+            and ${notBannedSQLFabric('cl', 'p', true)}
             and u.gone_status is null
         ) as c_likes,
         (select true from comment_likes cl
-          where cl.comment_id = comments.id
+          where cl.comment_id = c.id
             and cl.user_id = %L
         ) as has_own_like
-      from comments
-      where ${sqlIn('post_id', uniqPostsIds)} and (${hideCommentsSQL})
-    `,
+      from comments c join posts p on p.uid = c.post_id
+      where ${commentFilterSQL}`,
         viewerIntId,
       );
 
@@ -443,7 +415,7 @@ const timelinesPostsTrait = (superClass) =>
         : ``;
       const commentsSQL = `
       with comments as (${allCommentsSQL})
-      select ${commentFields.join(', ')}, count, c_likes, has_own_like from comments
+      select ${commentFields.join(', ')}, count, c_likes, has_own_like, hide_as_banned from comments
       ${foldCommentsSql}
       order by created_at, id
     `;
@@ -492,7 +464,7 @@ const timelinesPostsTrait = (superClass) =>
       }
 
       for (const comm of commentsData) {
-        if (!nobodyIsBanned && bannedUsersIds.includes(comm.user_id)) {
+        if (comm.hide_as_banned) {
           comm.user_id = null;
           comm.hide_type = Comment.HIDDEN_BANNED;
           comm.body = Comment.hiddenBody(Comment.HIDDEN_BANNED);